// server/prisma/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  ADMIN: 'ADMIN',
  USER: 'USER'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const CategoryUrlCode = {
  privateHouses: 'privateHouses',
  apartments: 'apartments',
  publicSpaces: 'publicSpaces'
} as const

export type CategoryUrlCode = (typeof CategoryUrlCode)[keyof typeof CategoryUrlCode]


export const ProjectImageType = {
  MAIN: 'MAIN',
  IMAGE: 'IMAGE',
  PLAN: 'PLAN',
  VIDEO: 'VIDEO'
} as const

export type ProjectImageType = (typeof ProjectImageType)[keyof typeof ProjectImageType]